Provide complete Etch-A-Sketch Project and UI
I wrote the entire HTML, CSS, and JavaScript of the project in one go.
The html contains the different elements such as the title, settings
buttons, color picker, and slider, and the sketch pad itself. The CSS
makes it look professional and clean. It especially provides hover
activity, proper element sizes, and color scheme. I wrote the script in
two versions. The first one was a remake of a previous successful
project while the second one, the file that is actually linked to the
webpage, is a self-written version of the first one. I wrote the second
script to completely understand how the script actually works. For
example, I created a grid with cells that contained event listeners and
proper width and height measurements as percentages. The buttons
acquired click listeners to change the current mode of the webpage. The
modes include color, rainbow, and eraser. Each mode decides the color of
the user's mouse paintbrush. The buttons also had an activate feature
that simply updates its css when clicked. The slider had its own input
listener where it resized and reloaded the grid based on the user's
input from 0 to 100. The clear button also simply reloaded the grid. All
in all, I came to a fundamental understanding of basic front end design.